FIXED hopefully fixed checkpoint save problems

+++ src/bert_model/PeptideBERTClasses/PeptideBertForRegression.py
@@ -102,7 +102,7 @@ class PeptideBertForRegression(BertModel):
         loss = None
         if labels is not None:
             loss_fct = nn.MSELoss()  # Use Binary Cross Entropy Loss
-            loss = loss_fct(logits.view(-1), labels.view(-1).float())
+            loss = loss_fct(logits.view(-1), labels.view(-1))
 
         if return_pooler_output:
             return pooled_output  # Return pooled output for visualization
